+++ examples/Example/BufferLogger/ExampleWithBufferLoggerInBufferLoggerBufferLoggerInBufferLogger.php
@@ -10,10 +10,11 @@ use Net\Bazzline\Component\ProxyLogger\Factory\BufferLoggerFactory;
 use Net\Bazzline\Component\ProxyLogger\Factory\LogRequestFactory;
 use Net\Bazzline\Component\ProxyLogger\Factory\LogRequestRuntimeBufferFactory;
 use Net\Bazzline\Component\ProxyLogger\OutputToConsoleLogger;
+use Net\Bazzline\Component\ProxyLogger\Proxy\BufferLoggerInterface;
 
 require_once __DIR__ . '/../../../vendor/autoload.php';
 
-Example::create()
+ExampleWithBufferLoggerInBufferLogger::create()
     ->setup()
     ->andRun();
 
@@ -24,17 +25,17 @@ Example::create()
  * @author stev leibelt <artodeto@arcor.de>
  * @since 2013-08-28
  */
-class Example
+class ExampleWithBufferLoggerInBufferLogger
 {
     /**
-     * @var \Net\Bazzline\Component\ProxyLogger\Proxy\BufferLogger
+     * @var \Net\Bazzline\Component\ProxyLogger\Proxy\BufferLoggerInterface
      * @author stev leibelt <artodeto@arcor.de>
      * @since 2013-08-28
      */
     private $bufferLogger;
 
     /**
-     * @return Example
+     * @return ExampleWithBufferLoggerInBufferLogger
      * @author stev leibelt <artodeto@arcor.de>
      * @since 2013-08-28
      */
@@ -57,7 +58,8 @@ class Example
         $bufferLoggerFactory->setLogRequestBufferFactory($logRequestBufferFactory);
         $logger = new OutputToConsoleLogger();
 
-        $this->bufferLogger = $bufferLoggerFactory->create($logger);
+        $innerBufferLogger = $bufferLoggerFactory->create($logger);
+        $this->bufferLogger = $bufferLoggerFactory->create($innerBufferLogger);
 
         return $this;
     }
@@ -74,8 +76,14 @@ class Example
         $this->bufferLogger->alert('Current line is ' . __LINE__);
         $this->bufferLogger->critical('Current line is ' . __LINE__);
         echo str_repeat('-', 40) . PHP_EOL;
-        echo 'Flush buffer' . PHP_EOL;
+        echo 'Flush buffer of outer logger' . PHP_EOL;
         $this->bufferLogger->flush();
+        echo 'Flush buffer of inner logger' . PHP_EOL;
+        foreach ($this->bufferLogger->getLoggers() as $logger) {
+            if ($logger instanceof BufferLoggerInterface) {
+                $logger->flush();
+            }
+        }
         echo str_repeat('-', 40) . PHP_EOL;
     }
 }