+++ documentation/VersionHistory.md
@@ -1,10 +1,10 @@
 # Version History
 
 * [next](https://github.com/stevleibelt/php_component_proxy_logger)
-    * add example how to add a proxy logger in a proxy logger
     * add output of example flush buffer trigger logger versus normal logger to readme
     * readme examples are also provided as try out code example
 * [1.1.0](https://github.com/stevleibelt/php_component_proxy_logger/tree/1.1.0) - not yet released
+    * add example how to add a proxy logger in a proxy logger
     * adapt factories, replace setter with factory setters where possible
     * adapt proxy logger factories and add awareInterfaces for buffer manipulator factories
     * create factories for buffer manipulator